move files to report-parsing module
