+++ apps/web/src/modules/report-parsing/utils/extractPdfText.ts
